feat: add provided meta to embeddings request
