+++ src/Text/Request.php
@@ -5,16 +5,16 @@ declare(strict_types=1);
 namespace EchoLabs\Prism\Text;
 
 use Closure;
+use EchoLabs\Prism\Concerns\AccessesProviderMeta;
 use EchoLabs\Prism\Concerns\ChecksSelf;
 use EchoLabs\Prism\Contracts\Message;
 use EchoLabs\Prism\Contracts\PrismRequest;
-use EchoLabs\Prism\Enums\Provider;
 use EchoLabs\Prism\Enums\ToolChoice;
 use EchoLabs\Prism\Tool;
 
 readonly class Request implements PrismRequest
 {
-    use ChecksSelf;
+    use AccessesProviderMeta, ChecksSelf;
 
     /**
      * @param  array<int, Message>  $messages
@@ -58,16 +58,5 @@ readonly class Request implements PrismRequest
             toolChoice: $this->toolChoice,
             providerMeta: $this->providerMeta,
         );
-    }
-
-    public function providerMeta(string|Provider $provider, string $valuePath = ''): mixed
-    {
-        $providerMeta = data_get(
-            $this->providerMeta,
-            is_string($provider) ? $provider : $provider->value,
-            []
-        );
-
-        return data_get($providerMeta, $valuePath, $providerMeta);
     }
 }
